Update and rename intakeCommand.java to FlywheelJoystick.java
joystick controlled flywheel
